refactor(separated scene & demo to own files)
